Make the smawk dependency optional
The smawk crate is now an optional (but default) dependency. It
remains a default dependency because it improves the result of the
word wrapping at very little cost for the user — conversely, it is now
an optional dependency so that people who don’t want to use the
optimal-fit wrapping algorithm can avoid the dependency.
