+++ Cargo.toml
@@ -25,10 +25,10 @@ name = "linear"
 harness = false
 
 [features]
-default = ["unicode-width"]
+default = ["unicode-width", "smawk"]
 
 [dependencies]
-smawk = "0.3"
+smawk = { version = "0.3", optional = true }
 unicode-width = { version= "0.1", optional = true }
 terminal_size = { version = "0.1", optional = true }
 hyphenation = { version = "0.8", optional = true, features = ["embed_en-us"] }
