+++ examples/interactive.rs
@@ -19,6 +19,7 @@ mod unix_only {
     use termion::raw::{IntoRawMode, RawTerminal};
     use termion::screen::AlternateScreen;
     use termion::{color, cursor, style};
+    #[cfg(feature = "smawk")]
     use textwrap::core::WrapAlgorithm::{FirstFit, OptimalFit};
     use textwrap::{wrap, HyphenSplitter, NoHyphenation, Options, WordSplitter};
 
@@ -102,15 +103,18 @@ mod unix_only {
         )?;
         left_row += 1;
 
-        write!(
-            stdout,
-            "{}- algorithm: {}{:?}{} (toggle with Ctrl-o)",
-            cursor::Goto(left_col, left_row),
-            style::Bold,
-            options.wrap_algorithm,
-            style::Reset,
-        )?;
-        left_row += 1;
+        #[cfg(feature = "smawk")]
+        {
+            write!(
+                stdout,
+                "{}- algorithm: {}{:?}{} (toggle with Ctrl-o)",
+                cursor::Goto(left_col, left_row),
+                style::Bold,
+                options.wrap_algorithm,
+                style::Reset,
+            )?;
+            left_row += 1;
+        }
 
         let now = std::time::Instant::now();
         let mut lines = wrap(text, options);
@@ -288,6 +292,7 @@ mod unix_only {
                 Key::Left => options.width = options.width.saturating_sub(1),
                 Key::Right => options.width = options.width.saturating_add(1),
                 Key::Ctrl('b') => options.break_words = !options.break_words,
+                #[cfg(feature = "smawk")]
                 Key::Ctrl('o') => {
                     options.wrap_algorithm = match options.wrap_algorithm {
                         OptimalFit => FirstFit,
